post-writer/src/main/java/pw/feed/postwriter/in/controller/PostController.java
package pw.feed.postwriter.in.controller;

import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import pw.feed.postwriter.model.post.Post;
import pw.feed.postwriter.service.PostService;

import java.util.List;

@RestController
@RequestMapping("/api/pw/posts")
@RequiredArgsConstructor
public class PostController {

    private final PostService postService;

    @PostMapping
    public ResponseEntity<Post> createPost(@RequestBody @Valid Post post) {
        return ResponseEntity
                .status(HttpStatus.CREATED)
                .body(postService.createPost(post));
    }

    @GetMapping
    public ResponseEntity<List<Post>> getAllPosts() {
        return ResponseEntity.ok(postService.getAllPosts());
    }

    @GetMapping("/{id}")
    public ResponseEntity<Post> getPostById(@PathVariable Long id) {
        return ResponseEntity.ok(postService.getPostById(id));
    }

    @PutMapping("/{id}")
    public ResponseEntity<Post> updatePost(@PathVariable Long id, @RequestBody @Valid Post post) {
        return ResponseEntity.ok(postService.updatePost(id, post));
    }
}
